requirements.txt
streamlit>=1.37
pandas>=2.2
openpyxl>=3.1
pyarrow>=16
numpy>=2.1
plotly>=5.17
xlrd>=2.0
streamlit-authenticator>=0.3.2
pyyaml>=6.0
psycopg2-binary>=2.9
sqlalchemy>=2.0
bcrypt>=4.0.0
cryptography>=41.0.0
extra-streamlit-components>=0.1.71
requests>=2.31.0
tomli>=2.0.0; python_version < '3.11'
scikit-learn>=1.3.0
xgboost>=2.0.0
